+++ internal/provider/fwprovider/provider.go
@@ -239,6 +239,10 @@ func (p *fwprovider) EphemeralResources(_ context.Context) []func() ephemeral.Ep
 		ephemeralsecrets.NewDBEphemeralSecretResource,
 		ephemeralsecrets.NewAzureStaticCredsEphemeralSecretResource,
 		ephemeralsecrets.NewAzureAccessCredentialsEphemeralResource,
+		ephemeralsecrets.NewGCPServiceAccountKeyEphemeralResource,
+		ephemeralsecrets.NewGCPOAuth2AccessTokenEphemeralResource,
+		ephemeralsecrets.NewAWSAccessCredentialsEphemeralSecretResource,
+		ephemeralsecrets.NewAWSStaticAccessCredentialsEphemeralSecretResource,
 	}
 
 }